chore: build ok + improve log, error messages

+++ run.sh
@@ -1,20 +1,38 @@
-#!/usr/bin/sh
+#!/bin/sh
 
 MODE=$1
 ENVIRONMENT=$2
 LANGUAGE="$(cut -d'-' -f1 <<<$ENVIRONMENT)"
 
+
+MAIN="\e[94m"
+ERR="\e[31m"
+_TEXT="\e[4m"
+ENDCOLOR="\e[0m"
+
+log() {
+    echo -e "$_TEXT$MAIN$@$ENDCOLOR"
+}
+
+logerr() {
+    echo -e "$_TEXT$ERR$@$ENDCOLOR"
+}
+
+to_underscore() {
+    echo "$1" | tr '-' '_'
+}
+
 usage() {
     echo "Usage: $0 <bin|container|dev> <environment>"
     echo "  bin: BUILD and RUN the binary"
-    echo "  container: RUN the container (build with ./build.sh)"
+    echo "  container: RUN the container (build before with ./build.sh)"
     echo "  dev: RUN the development server (hot reloading with Air, see CONTRIBUTING.md)"
 }
 
 # Check if the command exists 
 require() {
     command -v $1 >/dev/null 2>&1 || {
-        echo "Error: $1 is not installed"
+        logerr "Error: $1 is not installed"
         exit 1
     }
 }
@@ -23,11 +41,24 @@ require() {
 
 if [ $MODE = "dev" ]; then
     require air
-    air -build.cmd="go build -o ./tmp/main -tags $ENVIRONMENT,lib$LANGUAGE ."
+    require go
+    AIR_BUILD_CMD="go build -o ./tmp/main -tags $(to_underscore $ENVIRONMENT),lib$LANGUAGE ."
+    log "Running development server with \"$AIR_BUILD_CMD\""
+    air -build.cmd="$AIR_BUILD_CMD"
 elif [ $MODE = "bin" ]; then
+    require go
+    log "Running binary with tags $ENVIRONMENT and lib$LANGUAGE"
     go run -tags $ENVIRONMENT,lib$LANGUAGE .
 elif [ $MODE = "container" ]; then
     require docker
+
+    log "Building container with $ENVIRONMENT"
+    ./build.sh container -l $ENVIRONMENT
+
+    # If the build failed, exit
+    [ $? -ne 0 ] && exit 1
+
+    log "Running container with $ENVIRONMENT"
     docker run --rm --name srvexec-$ENVIRONMENT -p 8080:8080 srvexec:$ENVIRONMENT
 else
     echo 'Invalid mode!'
